performance optimizing by changing datatype
	modified:   alpha/alpha.ino
	modified:   mega/mega.ino
	modified:   mega/sensor.cpp
	modified:   mega/sensor.h

+++ mega/sensor.h
@@ -5,6 +5,6 @@
 #include "config.h"
 
 void sensorInit();
-void readSensor(float *sensorData);
+void readSensor(uint8_t *sensorData);
 
 #endif
